Add screensaver unit.
The `scrsave2` file is the actual unit, whereas `aftrdark` is a test app
for testing the various modules in the screensaver.

Originally there was a `scrsave` unit, and it had some different modules
but it was never used by any of my salvaged source code. So, I'll
include it eventually but for now I'm just committing dependencies, so
I'm only including `scrsave2` and its test file.