Added "Good_Traction" app.
A poem. Posted on the screen and then set to rotating colors. Why? I
don't know. Presumably it was a specific assignment by the instructor.
But it is an interesting exercising in using a "sleep" or "delay"
routine (here called `ChillOut`) as well as a procedure that can write
text with additional attributes, here used to assign color.