Create installers and runners for Turbo C++ and Turbo Pascal
The installers can get pointed at installer disk images and will help install the IDE in
the appropriate environment (Turbo C++ or Turbo Pascal).

The runner will help automatically open DosBox in the desired environment.

The installers and runners also make sure you have `dosbox-staging` installed as the
version of DosBox they'll be using. This is a more modern version of DosBox that has some
features that are required for the installers to work right, namely, handling of multiple
floppy disk images that can be swapped out during installation.