Add simple color display app in Pascal
This app simply cycles through all the colors by changing the text background to the next
color, while displaying that color number.

You might think that the arrow keys could be used to navigate up and down through the
numbers, but nope, it simply waits for any keystroke to increment the number. This is a
very common exploratory app that we kids wrote, giving us just enough to see the meaning
behind system values.