Add "Good Traction" poem app in Pascal
A poem. Posted on the screen and then set to rotating colors. Why? I have no idea!
Presumably this was a specific assignment by the instructor, which is ironic, since an
app that just outputs a chunk of static text was the making of my first several apps
written in BASIC as a little kid.

But it is an interesting exercise in using a "sleep" or "delay" routine (here called
`ChillOut`) as well as a procedure that can write text with additional attributes, here
used to assign color.

Several of these techniques were later incorporated into my personal libraries that I used
in future apps.