Add "Fake" app, a fake command line in Pascal
Original date: 3/24/96

This app just mimics a command-line, but will stop running with the `stop` command, or
will give a greeting with the `say hi` command.

Clearly this was just an experiment for future more sophisticated shells.

It's tempting to consider this a revision of DOScommander (DOSCMND.PAS), but it is so
different in structure that it has to be considered a separate program.

Nevertheless, I was indeed fascinated with the idea of making, essentially, my own shell
that sat on top of the system shell and added more functionality. If only I had been born
a little earlier, I could've been writing my own Linux shells.